+++ composer.json
@@ -1,15 +1,15 @@
 {
-    "name": "mnapoli/piwik-twig-extension",
+    "name": "mnapoli/matomo-twig-extension",
     "license": "MIT",
     "type": "library",
     "autoload": {
         "psr-4": {
-            "PiwikTwigExtension\\": "src/"
+            "MatomoTwigExtension\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Tests\\PiwikTwigExtension\\": "tests/"
+            "Tests\\MatomoTwigExtension\\": "tests/"
         }
     },
     "require": {
