+++ tests/NotConfiguredExtensionFixture/FunctionCallWithoutHost.test
@@ -1,8 +1,8 @@
 --TEST--
-Calling "piwik" function without parameters, leading to an error
+Calling "matomo" function without parameters, leading to an error
 --TEMPLATE--
-{{ piwik() }}
+{{ matomo() }}
 --DATA--
 return array();
 --EXPECT--
-Twig\Error\Error: Twig\Error\RuntimeError: An exception has been thrown during the rendering of a template ("No Piwik host was configured or given to generate the tracker code") in "index.twig" at line 2.
+Twig\Error\Error: Twig\Error\RuntimeError: An exception has been thrown during the rendering of a template ("No Matomo host was configured or given to generate the tracker code") in "index.twig" at line 2.
